Struct nth_cons_list::List

pub struct List<A> { /* fields omitted */ }

An immutable cons list.

Examples

A list can be created using nil and cons.

use nth_cons_list::{cons, nil};

// A list containing `1, 2, 3`
let list = cons(1, cons(2, cons(3, nil())));

Lists can be cheaply shared using .clone().

use std::fmt::Display;
use nth_cons_list::*;

fn print_list<A: Display>(list: List<A>) {
    println!("{}", list);
}

let list = cons(1, cons(2, cons(3, nil())));
print_list(list.clone());

Methods

impl<A> List<A>

pub fn head(&self) -> &A

Returns the first element of the list.

Panics

Panics if the list is empty.

pub fn tail(&self) -> List<A>

Returns a list containing all elements except the first.

Panics

Panics if the list is empty.

pub fn head_opt(&self) -> Option<&A>

Returns the first element of the list, or None if this list is empty.

pub fn tail_opt(&self) -> Option<List<A>>

Returns a list containing all elements except the first, or None if this list is empty.

pub fn is_empty(&self) -> bool

Tests whether this list is empty.

pub fn len(&self) -> usize

Returns the length of this list.

pub fn iter(&self) -> Iter<A>

Returns an iterator over the elements of this list.

pub fn reverse(&self) -> List<A> where
A: Clone,

Returns a list with the elements in reverse order.

pub fn from_double_ended_iter<I: DoubleEndedIterator<Item = A>>(
iter: I
) -> List<A>

Creates a new list from a DoubleEndedIterator.